requirements.txt
Django
Pillow
djangorestframework
# djangorestframework-jwt abondaned no latest django support
djangorestframework_simplejwt
pytest
pytest-django
arrow
requests
# psycopg2
psycopg2-binary
dj-static==0.0.6
django-countries==7.1
pyuca==1.2
django-cors-headers
